third_party/blink/web_tests/fast/url/idna2008-expected.txt

IDNA2008 handling in domain name labels.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

The PASS/FAIL results of this test are set to the behavior in IDNA2008.
PASS canonicalize('http://Bücher.de/') is 'http://xn--bcher-kva.de/'
PASS canonicalize('http://faß.de/') is 'http://xn--fa-hia.de/'
PASS canonicalize('http://βόλος.com/') is 'http://xn--nxasmm1c.com/'
PASS canonicalize('http://ශ්‍රී.com/') is 'http://xn--10cl1a0b660p.com/'
PASS canonicalize('http://نامه‌ای.com/') is 'http://xn--mgba3gch31f060k.com/'
PASS canonicalize('http://♥.net/') is 'http://xn--g6h.net/'
PASS canonicalize('http://͸.net/') is 'http://͸.net/'
PASS canonicalize('http://Ӏ.com/') is 'http://Ӏ.com/'
PASS canonicalize('http://㛼.com/') is 'http://㛼.com/'
PASS canonicalize('http://Ↄ.com/') is 'http://Ↄ.com/'
PASS canonicalize('http://look͏out.net/') is 'http://lookout.net/'
PASS canonicalize('http://gOoGle.com/') is 'http://google.com/'
PASS canonicalize('http://ড়.com/') is 'http://xn--15b8c.com/'
PASS canonicalize('http://ẞ.com/') is 'http://xn--zca.com/'
PASS canonicalize('http://-foo.bar.com/') is 'http://-foo.bar.com/'
PASS canonicalize('http://foo-.bar.com/') is 'http://foo-.bar.com/'
PASS canonicalize('http://ab--cd.com/') is 'http://ab--cd.com/'
PASS canonicalize('http://xn--0.com/') is 'http://xn--0.com/'
FAIL canonicalize('http://foò.bar.com/') should be http://foo%CC%80.bar.com/. Was http://xn--fo-3ja.bar.com/.
PASS successfullyParsed is true

TEST COMPLETE
